[STYLE] 패키지 구조 개편
- 관심사를 재정의하여 패키지 구조를 개편했습니다.
