+++ src/main/java/sopt/makers/authentication/support/code/base/FailureCode.java
@@ -1,0 +1,3 @@
+package sopt.makers.authentication.support.code.base;
+
+public interface FailureCode extends BaseCode {}
