--- src/main/java/sopt/makers/authentication/support/common/code/base/FailureCode.java
@@ -1,3 +1,0 @@
-package sopt.makers.authentication.support.common.code.base;
-
-public interface FailureCode extends BaseCode {}
