--- src/main/java/sopt/makers/authentication/support/common/code/base/SuccessCode.java
@@ -1,3 +1,0 @@
-package sopt.makers.authentication.support.common.code.base;
-
-public interface SuccessCode extends BaseCode {}
